+++ src-core/modules/common/delay_one_imag.h
@@ -5,7 +5,7 @@ class DelayOneImag
 {
 public:
     DelayOneImag();
-    void work(std::complex<float> *in, size_t length, std::complex<float> *out);
+    void work(std::complex<float> *in, int length, std::complex<float> *out);
 
 private:
     float lastSamp;
